# eliminate element groups and optimize some things