… bench/benchmarks/frame_duration.js
… js/data/buffer.js
… js/geo/transform.js
… js/render/draw_background.js
… js/render/draw_circle.js
… js/render/draw_collision_debug.js
… js/render/draw_debug.js
… js/render/draw_fill.js
… js/render/draw_line.js
… js/render/draw_raster.js
… js/render/draw_symbol.js
… js/render/painter.js
… js/render/vertex_array_object.js
… js/source/worker_tile.js
… shaders/circle.vertex.glsl
… shaders/icon.vertex.glsl
… shaders/sdf.vertex.glsl
… test/js/data/bucket.test.js
… test/js/source/worker_tile.test.js