- replaces element groups with buffer groups. Each buffer group has it's own buffers.
- cleans up and optimizes some things in drawSymbol
- makes frame-duration benchmark more independent of gpu performance

```
master:         11.7 ms
this branch:    8.6 ms
master:         11.3 ms
this branch:    8.3 ms
master:         11.9 ms
this branch:    9.3 ms

master average:         11.63 ms
this branch average:    8.7 ms
```

This branch (which includes the [unmerged vao changes](https://github.com/mapbox/mapbox-gl-js/pull/2483)) takes **75%** of the time it takes master to run through the render loop.

:eyes: @lucaswoj