@lucaswoj 2016-04-26
Just a couple syntactic-level comments. This looks great! Thank you for making our codebase more robust and faster.