@ansis 2016-04-26
@lucaswoj want to review the fixes?