@ansis 2016-04-27
merged into master